% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smet.R
\name{header}
\alias{header}
\title{retrieve header}
\usage{
header(smet)
}
\arguments{
\item{smet}{object of class smet}
}
\value{
named list of header information
}
\description{
retrieve header
}
\examples{
header(read_smet(system.file('examples/test.smet', package = 'smetlite')))
}
