% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smet.R
\name{read_smet}
\alias{read_smet}
\title{Read a SMET file into a smet dataframe}
\usage{
read_smet(filename)
}
\arguments{
\item{filename}{file or connection to be read by readr::read_table()}
}
\value{
dataframe with all
}
\description{
Read a SMET file into a smet dataframe
}
\examples{
read_smet(system.file('examples/test.smet', package = 'smetlite'))
}
