% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smet.R
\docType{class}
\name{smet-class}
\alias{smet-class}
\title{smet-class}
\description{
Class \code{smet}
}
\details{
class representing the SMET format file for weather station data
 \describe{
    \item{\code{signature}:}{A charachter string containing SMET signature}

    \item{\code{header}:}{Object of class \code{"list"} containing the Header Section,
     each key corresponds to an appropriately named element of the list.}
    \item{\code{data}:}{A \code{"data.frame"} containing the weather data values.
     The datetime field is called \code{"timestamp"} and stored in \code{\link{POSIXct}} format.}

	   \item{\code{file}:}{source. Either full name of the SMET file
	    or other source description like an url.}
 }

Detailed information about SMET format is described in \url{https://models.slf.ch/docserver/meteoio/SMET_specifications.pdf}.
Not all features of the SMET format are implemented:
* no binary data
* only numeric data are supported, except for timestamp
* no comments within the [DATA] section.
}
\note{
A \code{smet-class} object can be created from a SMET file via
function \code{\link{smet}}
}
\examples{

showClass("smet")

}
\references{
\url{https://models.slf.ch/docserver/meteoio/SMET_specifications.pdf}
}
\seealso{
\code{\link{smet}},\code{\link{print.smet}}

\code{\link{smet-class}}
}
\author{
Reinhold Koch
}
\keyword{classes}
