% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{smet_dict}
\alias{smet_dict}
\title{list of permissible time and measurement variables}
\format{
An object of class \code{list} of length 16.
}
\source{
<https://models.slf.ch/docserver/meteoio/SMET_specifications.pdf>
}
\usage{
smet_dict
}
\description{
list of permissible time and measurement variables
}
\keyword{datasets}
