% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_smet.R
\docType{data}
\name{test_smet}
\alias{test_smet}
\title{Test SMET data}
\format{
## `test_smet`
A data frame with 47 rows and 5 columns:
\describe{
  \item{timestamp}{Timestamp of measurement}
  \item{VW}{Velocity Wind, in m/s}
  \item{DW}{Direction Wind, in degrees, clockwise and north being zero degrees}
  \item{TA}{Temperature Air, in Kelvin}
  \item{HS}{Height Snow, in m}
}
}
\source{
<http://webgis.arpa.piemonte.it/webmeteo/meteo.php?CODTOT=001114901>
}
\usage{
test_smet
}
\description{
Artificial data derived from rifugio Vaccarone
}
\keyword{datasets}
